+++ CHANGELOG.md
@@ -52,6 +52,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Sanitize log messages (#1295)
 * Improve CosmosDB statement parser (#1282)
 * Make `ContractDefinitionService` get definition by its id (#1325)
+* Update the CodeQL version from v1 to v2
 * Update `slf4j-api` to `2.0.0-alpha7` (#1328)
 
 #### Removed
@@ -72,6 +73,8 @@ in the detailed section referring to by linking pull requests or issues.
 * Throw exception if `IdentityProviderKeyResolver` cannot get keys at startup (#1266)
 * Make all the services injectable (#1285)
 * Fix CosmosDB Integration tests (#1313)
+* Remove ContractDef from Cosmos DB cache when deleting (#1330)
+* Auto-upload of Cosmos stored procedures (#1338)
 
 ## [milestone-3] - 2022-04-08